+++ node_query.go
@@ -61,16 +61,9 @@ func buildFindChain(n *Node) error {
 	return buildFindChain(nextNode)
 }
 
-// findValueProperty finds the value for the given node using the property selector
-// information.
-func findValueProperty(n *Node, createIfNotExists bool) (reflect.Value, error) {
-	if !isValid(n.Previous.Value) {
-		return nilValue(), &UnexpectedPreviousNilValue{Selector: n.Previous.Selector.Current}
-	}
-
-	value := unwrapValue(n.Previous.Value)
-
-	if value.Kind() == reflect.Map {
+func findValuePropertyWork(n *Node, createIfNotExists bool, value reflect.Value) (reflect.Value, error) {
+	switch value.Kind() {
+	case reflect.Map:
 		for _, key := range value.MapKeys() {
 			if fmt.Sprint(key.Interface()) == n.Selector.Property {
 				return value.MapIndex(key), nil
@@ -80,9 +73,28 @@ func findValueProperty(n *Node, createIfNotExists bool) (reflect.Value, error) {
 			return nilValue(), nil
 		}
 		return nilValue(), &ValueNotFound{Selector: n.Selector.Current, PreviousValue: n.Previous.Value}
+
+	case reflect.Struct:
+		fieldV := value.FieldByName(n.Selector.Property)
+		if fieldV.IsValid() {
+			return fieldV, nil
+		}
+		return nilValue(), &ValueNotFound{Selector: n.Selector.Current, PreviousValue: n.Previous.Value}
+
+	case reflect.Ptr:
+		return findValuePropertyWork(n, createIfNotExists, derefValue(value))
 	}
 
 	return nilValue(), &UnsupportedTypeForSelector{Selector: n.Selector, Value: value}
+}
+
+// findValueProperty finds the value for the given node using the property selector
+// information.
+func findValueProperty(n *Node, createIfNotExists bool) (reflect.Value, error) {
+	if !isValid(n.Previous.Value) {
+		return nilValue(), &UnexpectedPreviousNilValue{Selector: n.Previous.Selector.Current}
+	}
+	return findValuePropertyWork(n, createIfNotExists, unwrapValue(n.Previous.Value))
 }
 
 // findValueIndex finds the value for the given node using the index selector
@@ -196,14 +208,7 @@ func findValueDynamic(n *Node, createIfNotExists bool) (reflect.Value, error) {
 	return nilValue(), &UnsupportedTypeForSelector{Selector: n.Selector, Value: value}
 }
 
-// findValueLength returns the length of the current node.
-func findValueLength(n *Node, createIfNotExists bool) (reflect.Value, error) {
-	if !isValid(n.Previous.Value) {
-		return nilValue(), &UnexpectedPreviousNilValue{Selector: n.Previous.Selector.Current}
-	}
-
-	value := unwrapValue(n.Previous.Value)
-
+func findValueLengthWork(n *Node, value reflect.Value) (reflect.Value, error) {
 	switch value.Kind() {
 	case reflect.Slice:
 		return reflect.ValueOf(value.Len()), nil
@@ -213,19 +218,26 @@ func findValueLength(n *Node, createIfNotExists bool) (reflect.Value, error) {
 
 	case reflect.String:
 		return reflect.ValueOf(value.Len()), nil
+
+	case reflect.Struct:
+		return reflect.ValueOf(value.NumField()), nil
+
+	case reflect.Ptr:
+		return findValueLengthWork(n, derefValue(value))
 	}
 
 	return nilValue(), &UnsupportedTypeForSelector{Selector: n.Selector, Value: value}
 }
 
-// findValueType returns the type of the current node.
-func findValueType(n *Node, createIfNotExists bool) (reflect.Value, error) {
+// findValueLength returns the length of the current node.
+func findValueLength(n *Node, createIfNotExists bool) (reflect.Value, error) {
 	if !isValid(n.Previous.Value) {
 		return nilValue(), &UnexpectedPreviousNilValue{Selector: n.Previous.Selector.Current}
 	}
+	return findValueLengthWork(n, unwrapValue(n.Previous.Value))
+}
 
-	value := unwrapValue(n.Previous.Value)
-
+func findValueTypeWork(n *Node, value reflect.Value) (reflect.Value, error) {
 	switch value.Kind() {
 	case reflect.Slice:
 		return reflect.ValueOf("array"), nil
@@ -244,9 +256,23 @@ func findValueType(n *Node, createIfNotExists bool) (reflect.Value, error) {
 
 	case reflect.Bool:
 		return reflect.ValueOf("bool"), nil
+
+	case reflect.Struct:
+		return reflect.ValueOf("struct"), nil
+
+	case reflect.Ptr:
+		return findValueTypeWork(n, derefValue(value))
 	}
 
 	return nilValue(), &UnsupportedTypeForSelector{Selector: n.Selector, Value: value}
+}
+
+// findValueType returns the type of the current node.
+func findValueType(n *Node, createIfNotExists bool) (reflect.Value, error) {
+	if !isValid(n.Previous.Value) {
+		return nilValue(), &UnexpectedPreviousNilValue{Selector: n.Previous.Selector.Current}
+	}
+	return findValueTypeWork(n, unwrapValue(n.Previous.Value))
 }
 
 // findValue finds the value for the given node.